+++ Database/DatabaseConfig.cs
@@ -2,8 +2,5 @@ namespace LabManager.Database;
 
 class DatabaseConfig
 {
-    public string ConnectionString
-    {
-        get => "Data Source=database.db";
-    }
+    public string ConnectionString { get => "Data Source=database.db";}  
 }
