+++ Database/DatabaseSetup.cs
@@ -5,6 +5,7 @@ namespace LabManager.Database;
 class DatabaseSetup
 {
     private DatabaseConfig databaseConfig;
+
     public DatabaseSetup(DatabaseConfig databaseConfig)
     {
         this.databaseConfig = databaseConfig;
@@ -14,7 +15,7 @@ class DatabaseSetup
 
     private void CreateTableComputer()
     {
-        var connection = new SqliteConnection("Data Source=database.db");
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
         connection.Open();
 
         var command = connection.CreateCommand();
