+++ Program.cs
@@ -1,12 +1,10 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
-
 new DatabaseSetup(databaseConfig);
-
-var computerRepository = new ComputerRepository(databaseConfig);
 
 //Routing
 var modelName = args[0];
@@ -14,6 +12,8 @@ var modelAction = args[1];
 
 if(modelName == "Computer")
 {
+    var computerRepository = new ComputerRepository(databaseConfig);
+
     if(modelAction == "List")
     {
         Console.WriteLine("Computer List");
@@ -25,21 +25,13 @@ if(modelName == "Computer")
 
     if(modelAction == "New")
     {
+        Console.WriteLine("Computer New");
         int id = Convert.ToInt32(args[2]);
         var ram = args[3];
         var processor = args[4];
-        
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-
-        command.ExecuteNonQuery();
-        connection.Close();
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
     }
 }
