+++ Repositories/ComputerRepository.cs
@@ -8,12 +8,16 @@ class ComputerRepository
 {
     private DatabaseConfig databaseConfig;
 
-    public ComputerRepository(DatabaseConfig databaseConfig) => this.databaseConfig = databaseConfig;
-   public  List<Computer> GetAll()
-   {
+    public ComputerRepository(DatabaseConfig databaseConfig)
+    {
+        this.databaseConfig = databaseConfig;
+    }
+    
+    public  List<Computer> GetAll()
+    {
        var computers = new List<Computer>();
 
-       var connection = new SqliteConnection("Data Source=database.db");
+       var connection = new SqliteConnection(databaseConfig.ConnectionString);
        connection.Open();
 
        var command = connection.CreateCommand();
@@ -29,5 +33,20 @@ class ComputerRepository
         connection.Close();
 
         return computers;
+    }
+
+    public void Save(Computer computer) 
+    {
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
     }
 }
